\name{Permutation1}
\alias{Permutation1}

\title{Permutation1}
\description{
This is a package containing different functions for making permutations. They all need a dataset (in dataframe format) with data that you want to test somehow with permutation tests. The functions also have the possibility of making plots showing the permuted null-distribution including the observed value. We recommend installing ggplot2 (part of the tidyverse as well) for this, if you want nice plots.
We have different functions depending on what statistical value you want to test, e.g. difference in mean.
}
\usage{
  Functions in this package (see more about the specific functions in the documentation about them):
  permutation_mean()
  permutation_median()
  permutation_lm()
  plot_permutation()
}
\author{
Adrián Gómez & Solveig Østergaard
}

